chore: Move utility
